Make sure only new miners claim to be new miners in BlocklHeaders.
Also updates:
- BlockHeaders to store the nickname as an uint32 as it's serialized as 4 unsigned bytes.
- SerializeBlockHeader test to use both keys and nicknames.

+++ src/Network/Serialize/Merit/ParseBlockHeader.nim
@@ -51,7 +51,7 @@ proc parseBlockHeader*(
                 headerSeq[1].toArgonHash(),
                 headerSeq[2].toHash(384),
                 headerSeq[3].toHash(384),
-                headerSeq[5].fromBinary(),
+                uint32(headerSeq[5].fromBinary()),
                 uint32(headerSeq[6].fromBinary()),
                 uint32(headerSeq[7].fromBinary()),
                 newBLSSignature(headerSeq[8])
